gha: extend matrix to include Windows and macOS, and lint on all
Combine "Lint" and "Test" into a single matrix so that linting runs
on all versions of Go and all platforms. Some linters apply rules
depending on Go version and we must avoid applying rules that are
not applicable to older versions of Go that are supported by this
module.

Combining "Lint" and "Test" allows re-using the common steps, such
as checking out the code and setting up Go; duration of tests and
linting on this repository is short, so running Lint and Test in
parallel does not gain much time.

This patch also reduces the test matrix to "oldest supported version
of Go (by this module), and currently supported versions of Go (current
stable, and previous stable)".

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>